refactor: move ~/entities/* to ~/*
