--- src/user/dto/login.dto.ts
@@ -1,14 +1,0 @@
-import { IsEmail, IsNotEmpty } from 'class-validator'
-
-import { ApiProperty } from '@nestjs/swagger'
-
-export class LoginUserDto {
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsEmail()
-  readonly email: string
-
-  @ApiProperty()
-  @IsNotEmpty()
-  readonly password: string
-}
